drivers: gpu: Limit GPU wakeups
When using an ioctl to wake up the GPU, it may still wake up
unnecessarily, even while not touching the screen, for example
when watching a video, or if certain events happen in top-app that
does not require the GPU to be woken, and so on.This may waste power
in the long run.

To mitigate this, it would be better if the GPU was woken up on ioctl
as well as on a touch event.This patch does exactly that.

(Referenced from the following commit:

https://github.com/ederekun/lazy_kernel_op5-t/commit/993124aff4ad4a8c7aabbbe1a9ac3153e2a52c79

as well as subsequent dynstune commits for the atomic_* function usages)

Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>