exfat: handle wrong stream entry size in exfat_readdir()
The compatibility between linux exfat and exfat of some camera company
was reported as a problem. In their exfat, if the number of files
exceeds any limit, the DataLength in stream entry of the directory is
not updated. There is a problem that some files created from camera
does not show in linux exfat. In linux exfat, if cpos becomes larger than
stream entry size, there is a check not to call exfat_readdir().
This patch check stream entry size only if it is ALLOC_NO_FAT_CHAIN and
add the check ensure that the number of dentry does not exceed max dentries
size(256 MB) to prevent the circular FAT chain issue.

Reported-by: Florian Cramer <flrncrmr@gmail.com>
Signed-off-by: Namjae Jeon <namjae.jeon@samsung.com>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>